#26 reorder tests
